updated make file

+++ Makefile
@@ -18,9 +18,11 @@ $(OUTPUT_DIR):
 	mkdir $(OUTPUT_DIR)
 
 %.o: %.c
-		$(CC) $(CCFLAGS) -c $< -o $(OUTPUT_DIR)/$@
+	$(CC) $(CCFLAGS) -c $< -o $(OUTPUT_DIR)/$@
 
-clean: | $(OUTPUT_DIR)
-	rd /s /q $(OUTPUT_DIR)
-
-
+clean:
+ifneq ($(OS),Windows_NT)
+	if [ -d $(OUTPUT_DIR) ]; then rm -rf $(OUTPUT_DIR); fi
+else
+	if exist $(OUTPUT_DIR) rd /s /q $(OUTPUT_DIR)
+endif
